Changes datatype of temperatures to float

+++ src/de/spreclib/model/enums/longtermstorage/LongTermStorageTemperature.java
@@ -15,7 +15,7 @@ public enum LongTermStorageTemperature {
     // Constructor for LIQUID_NITROGEN which has no Temperature in SPRECv2
   }
 
-  LongTermStorageTemperature(long lowerBound, long upperBound) {
+  LongTermStorageTemperature(float lowerBound, float upperBound) {
     // TODO Bounds überarbeiten
   }
 }
